requirements.txt
discord.py >= 2.3.2
python-dotenv
opengsq
asyncio
pycord
sqlite3
requests